backend/requirements-no-rust.txt
# Core FastAPI dependencies - no Rust compilation required
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Templates
jinja2==3.1.2

# Authentication - Pure Python implementations
PyJWT==2.8.0
# Use argon2 instead of bcrypt to avoid Rust compilation
argon2-cffi==23.1.0

# Web Scraping Dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# Scheduling and Monitoring (without Celery to avoid Redis compilation issues)
apscheduler==3.10.4

# Cloud Deployment Dependencies
gunicorn==21.2.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
